build: add initial ejected config
Via `ng eject --prod --aot --build-optimizer --extract-css --extract-licenses --vendor-chunk`.

The webpack config is being commited right now for reference, but will be removed at a future time.

Purposefully did not modify the Angular CLI config. The changes there were inconsequential to the scope of this project. The planned eject strategy we’ll take will not modify the core parts of the project.